converting 오류 수정

+++ recog_src/recog_preproc.py
@@ -4,7 +4,7 @@ def preprocess(input):
     input = input.resize([28,28])
     input = input.convert('L')
     input = convert2np(input)
-    input = float(input/255)
+    input = input/255
     print(input)
     input = input.reshape([1, 784])
     return input
